Revert "[SILOptimizer] Don't diagnose infinite recursion if a branch terminates (#19724)"
This reverts commit e94450e840009180e1e5ed90bd89112499bc0927.

rdar://45080912
