+++ include/swift/SIL/SILBasicBlock.h
@@ -370,11 +370,8 @@ public:
   /// no-return apply or builtin.
   bool isNoReturn() const;
 
-  /// Returns true if this block only contains a branch instruction.
+  /// Returns true if this instruction only contains a branch instruction.
   bool isTrampoline() const;
-
-  /// Returns true if this block traps without any side effects.
-  bool isProgramTerminationPoint() const;
 
   /// Returns true if it is legal to hoist instructions into this block.
   ///
