+++ lib/SIL/SILInstructions.cpp
@@ -1061,30 +1061,6 @@ bool TermInst::isFunctionExiting() const {
   llvm_unreachable("Unhandled TermKind in switch.");
 }
 
-bool TermInst::isProgramTerminating() const {
-  switch (getTermKind()) {
-  case TermKind::BranchInst:
-  case TermKind::CondBranchInst:
-  case TermKind::SwitchValueInst:
-  case TermKind::SwitchEnumInst:
-  case TermKind::SwitchEnumAddrInst:
-  case TermKind::DynamicMethodBranchInst:
-  case TermKind::CheckedCastBranchInst:
-  case TermKind::CheckedCastValueBranchInst:
-  case TermKind::CheckedCastAddrBranchInst:
-  case TermKind::TryApplyInst:
-  case TermKind::YieldInst:
-  case TermKind::ReturnInst:
-  case TermKind::ThrowInst:
-  case TermKind::UnwindInst:
-    return false;
-  case TermKind::UnreachableInst:
-    return true;
-  }
-
-  llvm_unreachable("Unhandled TermKind in switch.");
-}
-
 TermInst::SuccessorBlockArgumentsListTy
 TermInst::getSuccessorBlockArguments() const {
   function_ref<PhiArgumentArrayRef(const SILSuccessor &)> op;
