+++ lib/SILOptimizer/Analysis/SideEffectAnalysis.cpp
@@ -17,7 +17,6 @@
 #include "swift/SILOptimizer/Analysis/BasicCalleeAnalysis.h"
 #include "swift/SILOptimizer/Analysis/FunctionOrder.h"
 #include "swift/SILOptimizer/PassManager/PassManager.h"
-#include "swift/Strings.h"
 
 using namespace swift;
 
@@ -322,7 +321,7 @@ FunctionSideEffectFlags *FunctionSideEffects::getEffectsOn(SILValue Addr) {
 // Return true if the given function has defined effects that were successfully
 // recorded in this FunctionSideEffects object.
 bool FunctionSideEffects::setDefinedEffects(SILFunction *F) {
-  if (F->hasSemanticsAttr(SEMANTICS_PROGRAMTERMINATION_POINT)) {
+  if (F->hasSemanticsAttr("arc.programtermination_point")) {
     Traps = true;
     return true;
   }
