+++ stdlib/public/core/AssertCommon.swift
@@ -76,7 +76,6 @@ internal func _fatalErrorFlags() -> UInt32 {
 /// bloats code.
 @usableFromInline
 @inline(never)
-@_semantics("programtermination_point")
 internal func _assertionFailure(
   _ prefix: StaticString, _ message: StaticString,
   file: StaticString, line: UInt,
@@ -107,7 +106,6 @@ internal func _assertionFailure(
 /// bloats code.
 @usableFromInline
 @inline(never)
-@_semantics("programtermination_point")
 internal func _assertionFailure(
   _ prefix: StaticString, _ message: String,
   file: StaticString, line: UInt,
@@ -138,7 +136,6 @@ internal func _assertionFailure(
 /// bloats code.
 @usableFromInline
 @inline(never)
-@_semantics("programtermination_point")
 internal func _assertionFailure(
   _ prefix: StaticString, _ message: String,
   flags: UInt32
@@ -164,18 +161,27 @@ internal func _assertionFailure(
 /// bloats code.
 @usableFromInline
 @inline(never)
-@_semantics("programtermination_point")
+@_semantics("arc.programtermination_point")
 internal func _fatalErrorMessage(
   _ prefix: StaticString, _ message: StaticString,
   file: StaticString, line: UInt,
   flags: UInt32
 ) -> Never {
+  // This function breaks the infinite recursion detection pass by introducing
+  // an edge the pass doesn't look through.
+  func _withUTF8Buffer<R>(
+    _ string: StaticString,
+    _ body: (UnsafeBufferPointer<UInt8>) -> R
+  ) -> R {
+    return string.withUTF8Buffer(body)
+  }
+
 #if INTERNAL_CHECKS_ENABLED
-  prefix.withUTF8Buffer {
+  _withUTF8Buffer(prefix) {
     (prefix) in
-    message.withUTF8Buffer {
+    _withUTF8Buffer(message) {
       (message) in
-      file.withUTF8Buffer {
+      _withUTF8Buffer(file) {
         (file) in
         _swift_stdlib_reportFatalErrorInFile(
           prefix.baseAddress!, CInt(prefix.count),
@@ -186,9 +192,9 @@ internal func _fatalErrorMessage(
     }
   }
 #else
-  prefix.withUTF8Buffer {
+  _withUTF8Buffer(prefix) {
     (prefix) in
-    message.withUTF8Buffer {
+    _withUTF8Buffer(message) {
       (message) in
       _swift_stdlib_reportFatalError(
         prefix.baseAddress!, CInt(prefix.count),
@@ -265,7 +271,6 @@ func _undefined<T>(
 /// in release builds anyway (old apps that are run on new OSs).
 @inline(never)
 @usableFromInline // COMPILER_INTRINSIC
-@_semantics("programtermination_point")
 internal func _diagnoseUnexpectedEnumCaseValue<SwitchedValue, RawValue>(
   type: SwitchedValue.Type,
   rawValue: RawValue
@@ -283,7 +288,6 @@ internal func _diagnoseUnexpectedEnumCaseValue<SwitchedValue, RawValue>(
 /// in release builds anyway (old apps that are run on new OSs).
 @inline(never)
 @usableFromInline // COMPILER_INTRINSIC
-@_semantics("programtermination_point")
 internal func _diagnoseUnexpectedEnumCase<SwitchedValue>(
   type: SwitchedValue.Type
 ) -> Never {
